register contoller fix

+++ src/utils/cloudinary.js
@@ -10,14 +10,18 @@ cloudinary.config({
 const uploadToCloudinary=async(filePath)=>{
     try{
         if(!filePath) throw new Error("File path is required");
+        console.log("Uploading to cloudinary:", filePath);
         //uploading file to cloudinary
         const result = await cloudinary.uploader.upload(filePath,{
             resource_type:"auto",
         }) //auto-detect the file type(image,video etc)
         console.log("Cloudinary upload result:",result.url);
+        if (filePath && fs.existsSync(filePath)) fs.unlinkSync(filePath);//delete the file from local storage after upload
         return result;
     }catch(err){
-        fs.unlinkSync(filePath);//delete the file from local storage if error occurs
+        console.error("CLOUDINARY ERROR:", err.message);   // <-- IMPORTANT
+        console.error(err); 
+        if (filePath && fs.existsSync(filePath)) fs.unlinkSync(filePath);//delete the file from local storage if error occurs
         return null;
     }
 }
